# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forum DAF",
        "description": "Este projecto visa criar uma API REST para desenvolvimento de um site ou aplicação de\n          suporte, num formato de pergunta e respostas.<br />\n          Tem como principal objectivo testar a capacidade do aluno criar uma aplicação web, com\n          Symfony 4 e usando os padrões lecionados nas aulas da disciplina de “desenvolvimento de\n          aplicações com frameworks.",
        "contact": {
            "email": "silvam.filipe@gmail.com"
        },
        "version": "v0.1.0"
    },
    "paths": {
        "/questions/{questionId}/answers": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Adds a new answer",
                "operationId": "addAnswer",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AddAnswer"
                },
                "responses": {
                    "200": {
                        "description": "The newlly crated answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Answer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2.0-Token": []
                    }
                ]
            }
        },
        "/answers/{answerId}": {
            "get": {
                "tags": [
                    "Answers"
                ],
                "summary": "Retrieve the answer with provided ID",
                "description": "Returns an answer",
                "operationId": "getAnswer",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "id of answer to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Answer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2.0-Token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Answers"
                ],
                "summary": "Deletes an answer",
                "description": "Delete the answer that matches the provided answer ID",
                "operationId": "DeleteAnswer",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "id of answer to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The answer was deleted"
                    }
                },
                "security": [
                    {
                        "OAuth2.0-Token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Answers"
                ],
                "summary": "Update an answer",
                "operationId": "UpdateAnswer",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateAnswer"
                },
                "responses": {
                    "200": {
                        "description": "The updated answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateAnswer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2.0-Token": []
                    }
                ]
            }
        },
        "/answers/{answerId}/set-as-accepted": {
            "patch": {
                "tags": [
                    "Answers"
                ],
                "summary": "Mark an answer as the accepted one",
                "description": "Marks the answer as the accepted one for its question",
                "operationId": "AcceptAnswer",
                "parameters": [
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "id of answer to mark as correct",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated answer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Answer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2.0-Token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddAnswer": {
                "title": "AddAnswer",
                "properties": {
                    "description": {
                        "type": "string",
                        "example": "Tonight's dinner is lasagna!"
                    }
                },
                "type": "object"
            },
            "UpdateAnswer": {
                "title": "UpdateAnswer",
                "properties": {
                    "description": {
                        "type": "string",
                        "example": "Tonight's dinner is soup!"
                    }
                },
                "type": "object"
            },
...
            "Answer": {
                "properties": {
                    "answerId": {
                        "description": "Answer identifier",
                        "type": "string",
                        "example": "e1026e90-9b21-4b6d-b06e-9c592f7bdb82"
                    },
                    "questionId": {
                        "description": "Question identifier",
                        "type": "string",
                        "example": "e1026e90-9b21-4b6d-b06e-9c592f7bdb82"
                    },
                    "userId": {
                        "description": "User identifier",
                        "type": "string",
                        "example": "e1026e90-9b21-4b6d-b06e-9c592f7bdb82"
                    },
                    "description": {
                        "description": "Answer body",
                        "type": "string",
                        "example": "Yeah, off course I can help you, it is three o’clock."
                    },
                    "givenOn": {
                        "description": "Date and time answer was given",
                        "$ref": "#/components/schemas/DateTime"
                    },
                    "accepted": {
                        "description": "Flag the accepted/unaccepted answer state",
                        "type": "boolean",
                        "example": false
                    },
                    "voted": {
                        "description": "Flag the voted/non voted answer",
                        "type": "boolean",
                        "example": false
                    },
                    "lastEditedOn": {
                        "description": "Date and time answer was last edited",
                        "$ref": "#/components/schemas/DateTime"
                    },
                    "positiveVotes": {
                        "description": "Positive votes",
                        "type": "integer",
                        "example": 10
                    },
                    "negativeVotes": {
                        "description": "Negative votes",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "DateTime": {
                "properties": {
                    "date": {
                        "description": "Formatted date",
                        "type": "string",
                        "example": "2018-07-04 23:07:14.000000"
                    },
                    "timezone_type": {
                        "description": "Timezone type ID",
                        "type": "integer",
                        "format": "int32",
                        "example": "3"
                    },
                    "timezone": {
                        "description": "Timezone name",
                        "type": "string",
                        "example": "Atlantic/Azores"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
            "AddAnswer": {
                "description": "Object containing the very minimal information needded to create an answer",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/AddAnswer"
                        }
                    }
                }
            },
            "UpdateAnswer": {
                "description": "The updated answer",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateAnswer"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "OAuth2.0-Token": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "http://0.0.0.0:8080/auth/access-token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}